shifted motor_driver class and motor classes to the CommonComponents.py
