+++ CommonComponents.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 # -------------------------------------------------------------------------
 # class material
 # -------------------------------------------------------------------------
@@ -231,3 +233,496 @@ class nuts_and_bolts_dimensions:
         height = spec["height"]
 
         return [width_across_flats, height]
+
+
+# -------------------------------------------------------------------------
+# Motor Driver class
+# -------------------------------------------------------------------------
+class motor_driver:
+    def __init__(self, driver_name, motor_driver_data):
+        self.driver_name                          = driver_name
+        self.driver_upper_holes_dist_from_center  = motor_driver_data["driver_upper_holes_dist_from_center"]
+        self.driver_lower_holes_dist_from_center  = motor_driver_data["driver_lower_holes_dist_from_center"]
+        self.driver_side_holes_dist_from_center   = motor_driver_data["driver_side_holes_dist_from_center"]
+        self.driver_mount_holes_dia               = motor_driver_data["driver_mount_holes_dia"]
+        self.driver_mount_inserts_OD              = motor_driver_data["driver_mount_inserts_OD"]
+        self.driver_mount_thickness               = motor_driver_data["driver_mount_thickness"]
+        self.driver_mount_height                  = motor_driver_data["driver_mount_height"]
+
+        # self.print_vars()
+
+    def print_vars(self):
+        print("driver_name:", self.driver_name)
+        print("driver_upper_holes_dist_from_center: ", self.driver_upper_holes_dist_from_center)
+        print("driver_lower_holes_dist_from_center: ", self.driver_lower_holes_dist_from_center)
+        print("driver_side_holes_dist_from_center: ", self.driver_side_holes_dist_from_center)
+        print("driver_mount_holes_dia: ", self.driver_mount_holes_dia)
+        print("driver_mount_inserts_OD: ", self.driver_mount_inserts_OD)
+        print("driver_mount_thickness: ", self.driver_mount_thickness)
+        print("driver_mount_height: ", self.driver_mount_height)
+        print("---")
+
+
+# =========================================================================
+# Motor classes (5 distinct motor construction types)
+# =========================================================================
+
+# -------------------------------------------------------------------------
+# Framed Outrunner motor — used by SSPG, CPG, DSPG, WPG (from ActuatorAndGearbox.py)
+# -------------------------------------------------------------------------
+class motor_framed_outrunner:
+    def __init__(self, 
+                 maxMotorAngVelRPM     = 1190, # RPM
+                 maxMotorTorque        = 8.83, # Nm
+                 maxMotorPower         = 8.83 * 1190 * 2*np.pi/60, # W 
+                 motorMass             = 0.778, # KG
+                 motorDia              = 106.8, # mm
+                 motorLength           = 47.6,  # mm
+                 motor_mount_hole_PCD       = 32,
+                 motor_mount_hole_dia       = 4,
+                 motor_mount_hole_num       = 4,
+                 motor_output_hole_PCD      = 23,
+                 motor_output_hole_dia      = 4,
+                 motor_output_hole_num      = 4,
+                 wire_slot_dist_from_center = 30,
+                 wire_slot_length           = 10,
+                 wire_slot_radius           = 4,
+                 motorName             = "U12"):
+        
+        # Name of the motor
+        self.motorName                  = motorName
+
+        # Electrical and performance parameters of the motor
+        self.maxMotorAngVelRPM          = maxMotorAngVelRPM
+        self.maxMotorAngVelRadPerSec    = maxMotorAngVelRPM * (2 * np.pi / 60)
+        self.maxMotorTorque             = maxMotorTorque
+        self.maxMotorPower              = maxMotorPower
+        self.massKG                     = motorMass     # kg
+        
+        # Geometrical parameters of the motor
+        self.motorDiaMM                 = motorDia      # mm #TODO: make use of this parameter
+        self.motorLengthMM              = motorLength   # mm #TODO: make use of this parameter
+        self.motor_mount_hole_PCD       = motor_mount_hole_PCD 
+        self.motor_mount_hole_dia       = motor_mount_hole_dia
+        self.motor_mount_hole_num       = motor_mount_hole_num
+        self.motor_output_hole_PCD      = motor_output_hole_PCD 
+        self.motor_output_hole_dia      = motor_output_hole_dia
+        self.motor_output_hole_num      = motor_output_hole_num
+        self.wire_slot_dist_from_center = wire_slot_dist_from_center 
+        self.wire_slot_length           = wire_slot_length 
+        self.wire_slot_radius           = wire_slot_radius
+
+    def getMaxMotorAngVelRadPerSec(self): return self.maxMotorAngVelRadPerSec # Maximum motor angular velocity in rad/s
+    def getMaxMotorPower(self): return self.maxMotorPower # Maximum motor power in W
+    def getMaxMotorTorque(self): return self.maxMotorTorque # Maximum motor torque in Nm
+    def getMassKG(self): return self.massKG # Mass of the motor in kg
+    def getDiaMM(self): return self.motorDiaMM # dimension of the motor in MM
+    
+    def getLengthMM(self): return self.motorLengthMM
+
+    #dimensions of Stator
+    def getStatorIDMM(self): return self.motorStatorIDMM
+    def getStatorODMM(self): return self.motorStatorODMM
+    def getStatorHeight(self): return self.motorStatorHeightMM
+    
+    # Print the motor parameters
+    def printParameters(self):
+        print("Maximum motor angular velocity = ", self.maxMotorAngVelRPM, " RPM")
+        print("Maximum motor power = ", self.maxMotorPower, " W")
+        print("Maximum motor torque = ", self.maxMotorTorque, " Nm")
+        print("Maximum motor angular velocity = ", self.maxMotorAngVelRadPerSec, " rad/s")
+        print("Mass of the motor = ", self.massKG, " kg")
+        print ('Diameter of the motor = ', self.motorDiaMM, ' mm')
+        print( " Length of the motor = ", self.motorLengthMM, ' mm')
+        if (self.motorStatorIDMM != 0 and self.motorStatorODMM != 0 and self.motorStatorHeightMM !=0):
+            print (' Inner Diameter of Stator = ', self.motorStatorIDMM, ' mm')
+            print (' Outer Diameter of Stator = ', self.motorStatorODMM, ' mm')
+            print (' Height of Stator = ', self.motorStatorIDMM, ' mm')
+
+
+# -------------------------------------------------------------------------
+# Frameless Outrunner motor — ICPG version (from ActuatorandGearbox_ICPG.py)
+# -------------------------------------------------------------------------
+class motor_frameless_outrunner_mahi:
+    """
+    Stores all motor geometry and performance data.
+
+    Sections
+    ─────────
+    • Identification
+    • Performance parameters (Kv, torque, speed)
+    • Stator dimensions
+    • Rotor dimensions
+    • Bearing at motor–sun interface
+    • Compatibility aliases
+    """
+
+    def __init__(self,
+                 Kv                              = 55,
+                 maxContinuousCurrent            = 20,
+                 ratedVoltage                    = 48,
+                 power                           = 4560,
+                 massKG                          = 0.352,
+                 # ── Stator ──────────────────────────────────────────────
+                 statorODMM                      = 81,
+                 statorIDMM                      = 55,
+                 statorHeightMM                  = 23.8,
+                 statorMountingHolesPCDMM        = 63,
+                 motor_stator_extrusion_dia      = 88,
+                 motor_stator_extrusion_depth    = 2,
+                 # ── Rotor ───────────────────────────────────────────────
+                 rotorODMM                       = 92.6,
+                 rotorIDMM                       = 82.6,
+                 rotorHeightMM                   = 21.6,
+                 rotorBottomIDMM                 = 51,
+                 rotorBottomThicknessMM          = 2.6,
+                 rotorCSKHeadUpperDiaMM          = 8,
+                 rotorMountHolePCDMM             = 62,
+                 rotorMountHoleDiaMM             = 4,
+                 # ── Bearing at motor/sun interface ───────────────────────
+                 sun_bottom_casing_bearing_ID    = 50,
+                 sun_bottom_casing_bearing_OD    = 62,
+                 sun_bottom_casing_bearing_height= 6,
+                 sun_bottom_casing_bearing_massKG= 0.07,
+                 # ── Overall envelope ────────────────────────────────────
+                 motor_height                    = 36.2,
+                 motorName                       = "MotorR100"):
+
+        # ── Identification ─────────────────────────────────────────────────
+        self.motorName = motorName
+
+        # ── Performance parameters ─────────────────────────────────────────
+        self.Kv                   = Kv
+        self.maxContinuousCurrent = maxContinuousCurrent
+        self.ratedVoltage         = ratedVoltage
+        self.maxMotorPower        = power
+        self.massKG               = massKG
+
+        self.maxMotorAngVelRPM        = Kv * ratedVoltage
+        self.maxMotorAngVelRadPerSec  = self.maxMotorAngVelRPM * (2 * np.pi / 60)
+        self.maxMotorTorque           = maxContinuousCurrent / (Kv * 2 * np.pi / 60)
+
+        # ── Stator dimensions ──────────────────────────────────────────────
+        self.statorODMM                   = statorODMM
+        self.statorIDMM                   = statorIDMM
+        self.statorHeightMM               = statorHeightMM
+        self.statorMountingHolesPCDMM     = statorMountingHolesPCDMM
+        self.motor_stator_extrusion_dia   = motor_stator_extrusion_dia
+        self.motor_stator_extrusion_depth = motor_stator_extrusion_depth
+
+        # ── Rotor dimensions ───────────────────────────────────────────────
+        self.rotorODMM                = rotorODMM
+        self.rotorIDMM                = rotorIDMM
+        self.rotorHeightMM            = rotorHeightMM
+        self.rotorBottomIDMM          = rotorBottomIDMM
+        self.rotorBottomThicknessMM   = rotorBottomThicknessMM
+        self.rotorCSKHeadUpperDiaMM   = rotorCSKHeadUpperDiaMM
+        self.rotorMountHolePCDMM      = rotorMountHolePCDMM
+        self.rotorMountHoleDiaMM      = rotorMountHoleDiaMM
+
+        # ── Bearing at motor / sun interface ──────────────────────────────
+        self.sun_bottom_casing_bearing_ID     = sun_bottom_casing_bearing_ID
+        self.sun_bottom_casing_bearing_OD     = sun_bottom_casing_bearing_OD
+        self.sun_bottom_casing_bearing_height = sun_bottom_casing_bearing_height
+        self.sun_bottom_casing_bearing_massKG = sun_bottom_casing_bearing_massKG
+
+        # ── Overall envelope ───────────────────────────────────────────────
+        self.motor_height  = motor_height
+
+        # ── Compatibility aliases (used by Actuator class) ─────────────────
+        self.motorDiaMM    = rotorODMM
+        self.motorLengthMM = motor_height
+
+    # ── Getters ────────────────────────────────────────────────────────────
+    def getMaxMotorAngVelRadPerSec(self): return self.maxMotorAngVelRadPerSec
+    def getMaxMotorPower(self):           return self.maxMotorPower
+    def getMaxMotorTorque(self):          return self.maxMotorTorque
+    def getMassKG(self):                  return self.massKG
+    def getDiaMM(self):                   return self.motorDiaMM
+    def getLengthMM(self):                return self.motorLengthMM
+
+    # ─ Stator getters
+    def getStatorODMM(self):              return self.statorODMM
+    def getStatorIDMM(self):              return self.statorIDMM
+    def getStatorHeightMM(self):          return self.statorHeightMM
+
+    # ─ Rotor getters
+    def getRotorODMM(self):               return self.rotorODMM
+    def getRotorIDMM(self):               return self.rotorIDMM
+    def getRotorHeightMM(self):           return self.rotorHeightMM
+
+    # ── Print ──────────────────────────────────────────────────────────────
+    def printParameters(self):
+        print("Maximum motor angular velocity = ", round(self.maxMotorAngVelRPM, 2),         " RPM")
+        print("Maximum motor power            = ", self.maxMotorPower,                        " W")
+        print("Maximum continuous torque      = ", round(self.maxMotorTorque, 3),             " Nm")
+        print("Maximum angular velocity       = ", round(self.maxMotorAngVelRadPerSec, 2),    " rad/s")
+        print("Mass                           = ", self.massKG,                               " kg")
+        print("── Stator ──────────────────────────────────────────────")
+        print("  Outer Diameter = ", self.statorODMM, " mm")
+        print("  Inner Diameter = ", self.statorIDMM, " mm")
+        print("  Height         = ", self.statorHeightMM, " mm")
+        print("── Rotor ───────────────────────────────────────────────")
+        print("  Outer Diameter = ", self.rotorODMM, " mm")
+        print("  Inner Diameter = ", self.rotorIDMM, " mm")
+        print("  Height         = ", self.rotorHeightMM, " mm")
+
+
+# -------------------------------------------------------------------------
+# Frameless Outrunner motor — ISSPG version (from ISSPG_inside_gen_eq.py / ISSPG_compact_gen_eq.py)
+# -------------------------------------------------------------------------
+class motor_frameless_outrunner_suyash:
+    
+    def __init__(self,
+                 motor_OD                     = 92.6,
+                 stator_ID                    = 55,
+                 motor_rotor_base_thickness   = 2.6,
+                 motor_rotor_base_ID          = 51,
+                 rotor_height                 = 21.6,
+                 rotor_ID                     = 82.6,
+                 motor_stator_extrusion_depth = 1.5,
+                 motor_stator_extrusion_dia   = 68,
+                 stator_height                = 22,
+                 stator_OD                    = 81,
+                 stator_top_rotor_top_offset  = 4.8,
+                 stator_hole_dia              = 3,
+                 stator_hole_PCD              = 63,
+                 motor_rotor_hole_num         = 6,
+                 motor_rotor_hole_dia         = 4.2,
+                 motor_rotor_hole_PCD         = 62,
+                 motor_height                 = 26.4,
+                 maxMotorAngVelRPM            = 5040,  # RPM
+                 maxMotorTorque               = 1.3,   # Nm
+                 maxMotorPower                = 1.3 * 5040 * 2*np.pi/60,  # W
+                 motorMass                    = 0.265, # KG
+                 motorName                    = "RO100"):
+
+        self.motorName = motorName
+
+        # Physical geometry
+        self.motor_OD                     = motor_OD
+        self.stator_ID                    = stator_ID
+        self.motor_rotor_base_thickness   = motor_rotor_base_thickness
+        self.motor_rotor_base_ID          = motor_rotor_base_ID
+        self.rotor_height                 = rotor_height
+        self.rotor_ID                     = rotor_ID
+        self.motor_stator_extrusion_depth = motor_stator_extrusion_depth
+        self.motor_stator_extrusion_dia   = motor_stator_extrusion_dia
+        self.stator_height                = stator_height
+        self.stator_OD                    = stator_OD
+        self.stator_top_rotor_top_offset  = stator_top_rotor_top_offset
+        self.stator_hole_dia              = stator_hole_dia
+        self.stator_hole_PCD              = stator_hole_PCD
+        self.motor_rotor_hole_num         = motor_rotor_hole_num
+        self.motor_rotor_hole_dia         = motor_rotor_hole_dia
+        self.motor_rotor_hole_PCD         = motor_rotor_hole_PCD   
+        self.motor_height                 = motor_height
+
+        #Motor other param
+        self.maxMotorAngVelRPM            = maxMotorAngVelRPM
+        self.maxMotorAngVelRadPerSec      = maxMotorAngVelRPM * (2 * np.pi / 60)
+        self.maxMotorTorque               = maxMotorTorque
+        self.maxMotorPower                = maxMotorPower
+        self.massKG                       = motorMass     # kg
+
+    # ------------------------------------------------------------------
+    # Convenience getters (mirror reference code style)
+    # ------------------------------------------------------------------
+    
+    # Maximum motor angular velocity in rad/s
+    def getMaxMotorAngVelRadPerSec(self):
+        return self.maxMotorAngVelRadPerSec
+    
+    # Maximum motor power in W
+    def getMaxMotorPower(self):
+        return self.maxMotorPower
+    
+    # Maximum motor torque in Nm
+    def getMaxMotorTorque(self):
+        return self.maxMotorTorque
+    
+    # Mass of the motor in kg
+    def getMassKG(self):
+        return self.massKG
+
+    def getMotorODMM(self):
+       return self.motor_OD
+
+    def getMotorHeightMM(self): 
+        return self.motor_height
+
+    def getStatorIDMM(self):  
+        return self.stator_ID
+
+    def printParameters(self):
+        print(f"Motor: {self.motorName}")
+        print(f"  motor_OD           = {self.motor_OD} mm")
+        print(f"  motor_height       = {self.motor_height} mm")
+        print(f"  stator_ID          = {self.stator_ID} mm")
+        print(f"  stator_hole_PCD    = {self.stator_hole_PCD} mm")
+        print(f"  rotor_height       = {self.rotor_height} mm")
+
+
+# -------------------------------------------------------------------------
+# Frameless Inrunner motor — INSSPG version (from ActuatorandGearbox_INSSPG.py)
+# -------------------------------------------------------------------------
+class motor_frameless_inrunner_mahi:
+    def __init__(self,
+                 Kv                               = 55,
+                 maxContinuousCurrent             = 20,
+                 ratedVoltage                     = 48,
+                 power                            = 838,
+                 massKG                           = 0.50,
+                 Stator_ID                        = 57.0,
+                 Stator_OD                        = 104.0,
+                 stator_height                    = 24.5,
+                 Rotor_height                     = 15.0,
+                 Rotor_OD                         = 55.6,
+                 Rotor_ID                         = 45.0,
+                 rotor_mount_hole_dia             = 4.0,
+                 rotor_mount_hole_CSK_OD          = 8.0,
+                 rotor_mount_hole_CSK_head_hight  = 3.0,
+                 motorName                        = "RI100"):
+
+        self.motorName = motorName
+
+        self.Kv                   = Kv
+        self.maxContinuousCurrent = maxContinuousCurrent
+        self.ratedVoltage         = ratedVoltage
+        self.maxMotorPower        = power
+        self.massKG               = massKG
+
+        self.maxMotorAngVelRPM       = Kv * ratedVoltage
+        self.maxMotorAngVelRadPerSec = self.maxMotorAngVelRPM * (2 * np.pi / 60)
+        self.maxMotorTorque          = maxContinuousCurrent / (Kv * 2 * np.pi / 60)
+
+        self.Stator_ID                        = Stator_ID
+        self.Stator_OD                        = Stator_OD
+        self.stator_height                    = stator_height
+
+        self.Rotor_height                     = Rotor_height
+        self.Rotor_OD                         = Rotor_OD
+        self.Rotor_ID                         = Rotor_ID
+        self.rotor_mount_hole_dia             = rotor_mount_hole_dia
+        self.rotor_mount_hole_CSK_OD          = rotor_mount_hole_CSK_OD
+        self.rotor_mount_hole_CSK_head_hight  = rotor_mount_hole_CSK_head_hight
+
+        self.motorDiaMM    = Stator_OD
+        self.motorLengthMM = stator_height
+
+    def getMaxMotorAngVelRadPerSec(self): return self.maxMotorAngVelRadPerSec
+    def getMaxMotorPower(self):           return self.maxMotorPower
+    def getMaxMotorTorque(self):          return self.maxMotorTorque
+    def getMassKG(self):                  return self.massKG
+    def getDiaMM(self):                   return self.motorDiaMM
+    def getLengthMM(self):                return self.motorLengthMM
+    def getStatorODMM(self):              return self.Stator_OD
+    def getStatorIDMM(self):              return self.Stator_ID
+    def getStatorHeightMM(self):          return self.stator_height
+    def getRotorODMM(self):               return self.Rotor_OD
+    def getRotorIDMM(self):               return self.Rotor_ID
+    def getRotorHeightMM(self):           return self.Rotor_height
+
+    def printParameters(self):
+        print("Maximum motor angular velocity = ", round(self.maxMotorAngVelRPM, 2),         " RPM")
+        print("Maximum motor power            = ", self.maxMotorPower,                       " W")
+        print("Maximum continuous torque      = ", round(self.maxMotorTorque, 3),            " Nm")
+        print("Maximum angular velocity       = ", round(self.maxMotorAngVelRadPerSec, 2),    " rad/s")
+        print("Mass                           = ", self.massKG,                              " kg")
+
+
+# -------------------------------------------------------------------------
+# Frameless Inrunner motor — INCPG version (from InrunnerCPG_dependent_gen_eq.py / InrunnerCPG_independent_gen_eq.py)
+# -------------------------------------------------------------------------
+class motor_frameless_inrunner_suyash:
+    
+    def __init__(self,
+                 rotor_OD                     = 55.6,
+                 stator_ID                    = 57,
+                 rotor_height                 = 15,
+                 rotor_ID                     = 45,
+                 stator_height                = 24.5,
+                 stator_OD                    = 104,
+                 stator_hole_dia              = 3,
+                 stator_top_height            = 7,
+                 stator_mid_height            = 13,
+                 stator_bottom_height         = 4.5,
+                 stator_inside_OD             = 101,
+                 stator_hole_num              = 4,
+                 stator_inside_ID             = 58,
+                 maxMotorAngVelRPM            = 5040,  # RPM
+                 maxMotorTorque               = 1.3,   # Nm
+                 maxMotorPower                = 1.3 * 5040 * 2*np.pi/60,  # W
+                 motorMass                    = 0.265, # KG
+                 motorName                    = "RO100"):
+
+        self.motorName = motorName
+
+        # Physical geometry
+        self.rotor_OD                     = rotor_OD
+        self.stator_ID                    = stator_ID
+        self.rotor_height                 = rotor_height
+        self.rotor_ID                     = rotor_ID
+        self.stator_height                = stator_height
+        self.stator_OD                    = stator_OD
+        self.stator_hole_dia              = stator_hole_dia
+        self.stator_top_height            = stator_top_height
+        self.stator_mid_height            = stator_mid_height
+        self.stator_bottom_height         = stator_bottom_height
+        self.stator_inside_OD             = stator_inside_OD
+        self.stator_hole_num              = stator_hole_num
+        self.stator_inside_ID             = stator_inside_ID
+        self.motorDiaMM                   =  self.stator_OD
+        self.motorLengthMM                = self.stator_top_height + self.stator_mid_height + self.stator_bottom_height 
+
+        #Motor other param
+        self.maxMotorAngVelRPM            = maxMotorAngVelRPM
+        self.maxMotorAngVelRadPerSec      = maxMotorAngVelRPM * (2 * np.pi / 60)
+        self.maxMotorTorque               = maxMotorTorque
+        self.maxMotorPower                = maxMotorPower
+        self.massKG                       = motorMass     # kg
+
+    # ------------------------------------------------------------------
+    # Convenience getters (mirror reference code style)
+    # ------------------------------------------------------------------
+    
+    # Maximum motor angular velocity in rad/s
+    def getMaxMotorAngVelRadPerSec(self):
+        return self.maxMotorAngVelRadPerSec
+    
+    # Maximum motor power in W
+    def getMaxMotorPower(self):
+        return self.maxMotorPower
+    
+    # Maximum motor torque in Nm
+    def getMaxMotorTorque(self):
+        return self.maxMotorTorque
+    
+    # Mass of the motor in kg
+    def getMassKG(self):
+        return self.massKG
+
+    def getDiaMM(self):
+       return self.motorDiaMM
+
+    def getLengthMM(self): 
+        return self.motorLengthMM
+
+    def getRotorIDMM(self):  
+        return self.rotor_ID
+
+    def printParameters(self):
+        print(f"Motor: {self.motorName}")
+        print(f"  rotor_OD              = {self.rotor_OD} mm")
+        print(f"  stator_ID             = {self.stator_ID} mm")
+        print(f"  rotor_height          = {self.rotor_height} mm")
+        print(f"  rotor_ID              = {self.rotor_ID} mm")
+        print(f"  stator_height         = {self.stator_height} mm")
+        print(f"  stator_OD             = {self.stator_OD} mm")
+        print(f"  stator_hole_dia       = {self.stator_hole_dia} mm")
+        print(f"  stator_top_height     = {self.stator_top_height} mm")
+        print(f"  stator_mid_height     = {self.stator_mid_height} mm")
+        print(f"  stator_bottom_height  = {self.stator_bottom_height} mm")
+        print(f"  stator_inside_OD      = {self.stator_inside_OD} mm")
+        print(f"  stator_hole_num       = {self.stator_hole_num}")
+        print(f"  stator_inside_ID      = {self.stator_inside_ID} mm")
