+++ ISSPG_compact_gen_eq.py
@@ -5,122 +5,7 @@ import numpy as np
 import sys
 import time
 
-from CommonComponents import bearings_discrete, nuts_and_bolts_dimensions
-
-class motor:
-    
-    def __init__(self,
-                 motor_OD                     = 92.6,
-                 stator_ID                    = 55,
-                 motor_rotor_base_thickness   = 2.6,
-                 motor_rotor_base_ID          = 51,
-                 rotor_height                 = 21.6,
-                 rotor_ID                     = 82.6,
-                 motor_stator_extrusion_depth = 1.5,
-                 motor_stator_extrusion_dia   = 68,
-                 stator_height                = 22,
-                 stator_OD                    = 81,
-                 stator_top_rotor_top_offset  = 4.8,
-                 stator_hole_dia              = 3,
-                 stator_hole_PCD              = 63,
-                 motor_rotor_hole_num         = 6,
-                 motor_rotor_hole_dia         = 4.2,
-                 motor_rotor_hole_PCD         = 62,
-                 motor_height                 = 26.4,
-                 maxMotorAngVelRPM            = 5040,  # RPM
-                 maxMotorTorque               = 1.3,   # Nm
-                 maxMotorPower                = 1.3 * 5040 * 2*np.pi/60,  # W
-                 motorMass                    = 0.265, # KG
-                 motorName                    = "RO100"):
-
-        self.motorName = motorName
-
-        # Physical geometry
-        self.motor_OD                     = motor_OD
-        self.stator_ID                    = stator_ID
-        self.motor_rotor_base_thickness   = motor_rotor_base_thickness
-        self.motor_rotor_base_ID          = motor_rotor_base_ID
-        self.rotor_height                 = rotor_height
-        self.rotor_ID                     = rotor_ID
-        self.motor_stator_extrusion_depth = motor_stator_extrusion_depth
-        self.motor_stator_extrusion_dia   = motor_stator_extrusion_dia
-        self.stator_height                = stator_height
-        self.stator_OD                    = stator_OD
-        self.stator_top_rotor_top_offset  = stator_top_rotor_top_offset
-        self.stator_hole_dia              = stator_hole_dia
-        self.stator_hole_PCD              = stator_hole_PCD
-        self.motor_rotor_hole_num         = motor_rotor_hole_num
-        self.motor_rotor_hole_dia         = motor_rotor_hole_dia
-        self.motor_rotor_hole_PCD         = motor_rotor_hole_PCD   
-        self.motor_height                 = motor_height
-
-        #Motor other param
-        self.maxMotorAngVelRPM            = maxMotorAngVelRPM
-        self.maxMotorAngVelRadPerSec      = maxMotorAngVelRPM * (2 * np.pi / 60)
-        self.maxMotorTorque               = maxMotorTorque
-        self.maxMotorPower                = maxMotorPower
-        self.massKG                       = motorMass     # kg
-
-    # ------------------------------------------------------------------
-    # Convenience getters (mirror reference code style)
-    # ------------------------------------------------------------------
-    
-    # Maximum motor angular velocity in rad/s
-    def getMaxMotorAngVelRadPerSec(self):
-        return self.maxMotorAngVelRadPerSec
-    
-    # Maximum motor power in W
-    def getMaxMotorPower(self):
-        return self.maxMotorPower
-    
-    # Maximum motor torque in Nm
-    def getMaxMotorTorque(self):
-        return self.maxMotorTorque
-    
-    # Mass of the motor in kg
-    def getMassKG(self):
-        return self.massKG
-
-    def getMotorODMM(self):
-       return self.motor_OD
-
-    def getMotorHeightMM(self): 
-        return self.motor_height
-
-    def getStatorIDMM(self):  
-        return self.stator_ID
-
-    def printParameters(self):
-        print(f"Motor: {self.motorName}")
-        print(f"  motor_OD           = {self.motor_OD} mm")
-        print(f"  motor_height       = {self.motor_height} mm")
-        print(f"  stator_ID          = {self.stator_ID} mm")
-        print(f"  stator_hole_PCD    = {self.stator_hole_PCD} mm")
-        print(f"  rotor_height       = {self.rotor_height} mm")
-
-class motor_driver:
-    def __init__(self, driver_name, motor_driver_data):
-        self.driver_name                          = driver_name
-        self.driver_upper_holes_dist_from_center  = motor_driver_data["driver_upper_holes_dist_from_center"]
-        self.driver_lower_holes_dist_from_center  = motor_driver_data["driver_lower_holes_dist_from_center"]
-        self.driver_side_holes_dist_from_center   = motor_driver_data["driver_side_holes_dist_from_center"]
-        self.driver_mount_holes_dia               = motor_driver_data["driver_mount_holes_dia"]
-        self.driver_mount_inserts_OD              = motor_driver_data["driver_mount_inserts_OD"]
-        self.driver_mount_thickness               = motor_driver_data["driver_mount_thickness"]
-        self.driver_mount_height                  = motor_driver_data["driver_mount_height"]
-
-        # self.print_vars()
-    
-    def print_vars(self):
-        print("driver_name:", self.driver_name)
-        print("driver_upper_holes_dist_from_center: ", self.driver_upper_holes_dist_from_center)
-        print("driver_lower_holes_dist_from_center: ", self.driver_lower_holes_dist_from_center)
-        print("driver_side_holes_dist_from_center: ", self.driver_side_holes_dist_from_center)
-        print("driver_mount_holes_dia: ", self.driver_mount_holes_dia)
-        print("driver_mount_inserts_OD: ", self.driver_mount_inserts_OD)
-        print("driver_mount_thickness: ", self.driver_mount_thickness)
-        print("driver_mount_height: ", self.driver_mount_height)
-        print("---")
+from CommonComponents import bearings_discrete, nuts_and_bolts_dimensions, motor_driver, motor_frameless_outrunner_suyash as motor
 
 class internalsingleStagePlanetaryGearbox:
     def __init__(self,
