+++ InrunnerCPG_dependent_gen_eq.py
@@ -5,128 +5,8 @@ import numpy as np
 import sys
 import time
 
-from CommonComponents import material, bearings_discrete, nuts_and_bolts_dimensions
+from CommonComponents import material, bearings_discrete, nuts_and_bolts_dimensions, motor_driver, motor_frameless_inrunner_suyash as motor
 
-class motor:
-    
-    def __init__(self,
-                 rotor_OD                     = 55.6,
-                 stator_ID                    = 57,
-                 rotor_height                 = 15,
-                 rotor_ID                     = 45,
-                 stator_height                = 24.5,
-                 stator_OD                    = 104,
-                 stator_hole_dia              = 3,
-                 stator_top_height            = 7,
-                 stator_mid_height            = 13,
-                 stator_bottom_height         = 4.5,
-                 stator_inside_OD             = 101,
-                 stator_hole_num              = 4,
-                 stator_inside_ID             = 58,
-                 maxMotorAngVelRPM            = 5040,  # RPM
-                 maxMotorTorque               = 1.3,   # Nm
-                 maxMotorPower                = 1.3 * 5040 * 2*np.pi/60,  # W
-                 motorMass                    = 0.265, # KG
-                 motorName                    = "RO100"):
-
-        self.motorName = motorName
-
-        # Physical geometry
-        self.rotor_OD                     = rotor_OD
-        self.stator_ID                    = stator_ID
-        self.rotor_height                 = rotor_height
-        self.rotor_ID                     = rotor_ID
-        self.stator_height                = stator_height
-        self.stator_OD                    = stator_OD
-        self.stator_hole_dia              = stator_hole_dia
-        self.stator_top_height            = stator_top_height
-        self.stator_mid_height            = stator_mid_height
-        self.stator_bottom_height         = stator_bottom_height
-        self.stator_inside_OD             = stator_inside_OD
-        self.stator_hole_num              = stator_hole_num
-        self.stator_inside_ID             = stator_inside_ID
-        self.motorDiaMM                   =  self.stator_OD
-        self.motorLengthMM                = self.stator_top_height + self.stator_mid_height + self.stator_bottom_height 
-
-        #Motor other param
-        self.maxMotorAngVelRPM            = maxMotorAngVelRPM
-        self.maxMotorAngVelRadPerSec      = maxMotorAngVelRPM * (2 * np.pi / 60)
-        self.maxMotorTorque               = maxMotorTorque
-        self.maxMotorPower                = maxMotorPower
-        self.massKG                       = motorMass     # kg
-
-    # ------------------------------------------------------------------
-    # Convenience getters (mirror reference code style)
-    # ------------------------------------------------------------------
-    
-    # Maximum motor angular velocity in rad/s
-    def getMaxMotorAngVelRadPerSec(self):
-        return self.maxMotorAngVelRadPerSec
-    
-    # Maximum motor power in W
-    def getMaxMotorPower(self):
-        return self.maxMotorPower
-    
-    # Maximum motor torque in Nm
-    def getMaxMotorTorque(self):
-        return self.maxMotorTorque
-    
-    # Mass of the motor in kg
-    def getMassKG(self):
-        return self.massKG
-
-    def getDiaMM(self):
-       return self.motorDiaMM
-
-    def getLengthMM(self): 
-        return self.motorLengthMM
-
-    def getRotorIDMM(self):  
-        return self.rotor_ID
-
-    def printParameters(self):
-        print(f"Motor: {self.motorName}")
-        print(f"  rotor_OD              = {self.rotor_OD} mm")
-        print(f"  stator_ID             = {self.stator_ID} mm")
-        print(f"  rotor_height          = {self.rotor_height} mm")
-        print(f"  rotor_ID              = {self.rotor_ID} mm")
-        print(f"  stator_height         = {self.stator_height} mm")
-        print(f"  stator_OD             = {self.stator_OD} mm")
-        print(f"  stator_hole_dia       = {self.stator_hole_dia} mm")
-        print(f"  stator_top_height     = {self.stator_top_height} mm")
-        print(f"  stator_mid_height     = {self.stator_mid_height} mm")
-        print(f"  stator_bottom_height  = {self.stator_bottom_height} mm")
-        print(f"  stator_inside_OD      = {self.stator_inside_OD} mm")
-        print(f"  stator_hole_num       = {self.stator_hole_num}")
-        print(f"  stator_inside_ID      = {self.stator_inside_ID} mm")
-
-class motor_driver:
-    def __init__(self, driver_name, motor_driver_data):
-        self.driver_name                          = driver_name
-        self.driver_upper_holes_dist_from_center  = motor_driver_data["driver_upper_holes_dist_from_center"]
-        self.driver_lower_holes_dist_from_center  = motor_driver_data["driver_lower_holes_dist_from_center"]
-        self.driver_side_holes_dist_from_center   = motor_driver_data["driver_side_holes_dist_from_center"]
-        self.driver_mount_holes_dia               = motor_driver_data["driver_mount_holes_dia"]
-        self.driver_mount_inserts_OD              = motor_driver_data["driver_mount_inserts_OD"]
-        self.driver_mount_thickness               = motor_driver_data["driver_mount_thickness"]
-        self.driver_mount_height                  = motor_driver_data["driver_mount_height"]
-
-        # self.print_vars()
-    
-    def print_vars(self):
-        print("driver_name:", self.driver_name)
-        print("driver_upper_holes_dist_from_center: ", self.driver_upper_holes_dist_from_center)
-        print("driver_lower_holes_dist_from_center: ", self.driver_lower_holes_dist_from_center)
-        print("driver_side_holes_dist_from_center: ", self.driver_side_holes_dist_from_center)
-        print("driver_mount_holes_dia: ", self.driver_mount_holes_dia)
-        print("driver_mount_inserts_OD: ", self.driver_mount_inserts_OD)
-        print("driver_mount_thickness: ", self.driver_mount_thickness)
-        print("driver_mount_height: ", self.driver_mount_height)
-        print("---")
-
-#-------------------------------------------------------------------------
-# Compound Planetary Gearbox
-#-------------------------------------------------------------------------
 class inrunnerCompoundPlanetaryGearbox:
     def __init__(self,
                  design_parameters,
